requirements.txt
arxiv
pypdf2
openai
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
chromadb
reportlab
python-dotenv
numpy
